test: infrastructure improvements
Closes #367
